Sort icons before packing
Currently, the output of sprite zero is dependent on the order of the
icons array passed in. This can lead to an unintentional mismatch of json
and png unless the input order is consistent between calls.

Now, the icons array is sorted by height ascending and then by id before
packing to ensure that the same icon set will always produce the same
sprite sheet.

The test fixtures can be regenerated by `npm run regenerate-fixtures`.